refactor: UserRequestDto에 검증 애너테이션 추가

+++ src/main/resources/data.sql
@@ -5,7 +5,7 @@ DROP TABLE IF EXISTS car CASCADE;
 
 CREATE TABLE game_result
 (
-    id        BIGINT      NOT NULL AUTO_INCREMENT,
+    id        BIGINT   NOT NULL AUTO_INCREMENT,
     try_count INT      NOT NULL,
     played_at DATETIME NOT NULL default current_timestamp,
     PRIMARY KEY (id)
@@ -13,7 +13,7 @@ CREATE TABLE game_result
 
 CREATE TABLE car
 (
-    id             INT          NOT NULL AUTO_INCREMENT,
+    id             BIGINT       NOT NULL AUTO_INCREMENT,
     player_name    VARCHAR(255) NOT NULL,
     final_position INT          NOT NULL,
     is_winner      BOOLEAN      NOT NULL,
